extended config removed from separate file
